Move MavenCompilerUtils to utils package.
